rename files to lowercase
