+++ src/components/App/app.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import propTypes from 'prop-types';
-import Main from '../Main/Main';
+import Main from '../Main/main';
 
 const onCardClick = () => {
   // console.log(`Кликнули по фильму ${name}`);
