+++ src/index.jsx
@@ -1,5 +1,5 @@
 import ReactDOM from 'react-dom';
-import App from './components/App/App';
+import App from './components/App/app';
 
 const films = [
   `Fantastic Beasts`,
